requirement.txt
fastapi
uvicorn
git+https://github.com/huggingface/diffusers.git
transformers
accelerate
torch
Pillow
openai
python-dotenv
httpx